@@ -1,5 +1,5 @@
 {
-  "name": "proyecto",
+  "name": "maricaya",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -10,18 +10,20 @@
   "dependencies": {
     "axios": "^0.25.0",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11",
-    "vue-router": "^3.5.3"
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0-0",
+    "vue-sweetalert2": "^5.0.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-router": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "tailwindcss": "^3.0.18",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-vue": "^7.0.0",
+    "tailwindcss": "^3.0.18"
   },
   "eslintConfig": {
     "root": true,
@@ -29,7 +31,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
