@@ -1,10 +1,8 @@
-import Vue from 'vue'
-import App from './App.vue'
-import router from "./router";
+import { createApp } from 'vue';
+import App from './App.vue';
+import router from './router';
+import VueSweetalert2 from 'vue-sweetalert2';
+import 'sweetalert2/dist/sweetalert2.min.css';
+import "../css/output.css"
 
-Vue.config.productionTip = false
-
-new Vue({
-  router,
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).use(router).use(VueSweetalert2).mount('#app');
