@@ -1,6 +1,6 @@
 <template>
-  <div class="w-full bg-indigo-200">
-    <table class="w-full">
+  <div class="w-full flex justify-center">
+    <table class="w-11/12 mt-10">
       <thead>
         <tr class="bg-grey-100 border">
           <th class="text-left border p-2 border-black">ID</th>
@@ -24,7 +24,7 @@
         <td class="text-black border border-black">
           <div class="flex flex-row justify-center items-center gap-x-2">
             <button
-              v-on:click="deleteUser(user.id)"
+              v-on:click="showDeleteAlert(user.id)"
               onclick=""
               class="
                 bg-red-500
@@ -38,7 +38,7 @@
               Eliminar
             </button>
             <button
-              onclick=""
+              v-on:click="showAlert()"
               class="
                 bg-green-500
                 px-2
@@ -59,6 +59,7 @@
 </template>
 
 <script>
+//import Swal from 'vue-sweetalert2';
 import axios from "axios";
 export default {
   name: "ListarUsuario",
@@ -91,6 +92,27 @@ export default {
         }
       );
       this.getUsers();
+    },
+    showDeleteAlert(id) {
+      // Use sweetalert2
+      this.$swal({
+        title: "¿Estás seguro de hacerlo?",
+        text: "No podrás revertirlo!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Sí, quiero borrarlo!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.deleteUser(id);
+          this.$swal("Eliminado!", "El registro ha sido borrado", "success");
+        }
+      });
+    },
+    showAlert() {
+      // Use sweetalert2
+      this.$swal('Hello Vue world!!!');
     },
   },
 };
